adds styling to several pages, adds ProfileUpdate view and ProfileForm form
