+++ main_app/forms.py
@@ -1,5 +1,5 @@
 from django.forms import ModelForm
-from .models import Meal, MealFoodItem, BodyData
+from .models import Meal, MealFoodItem, BodyData, Food, Profile
 
 class MealForm(ModelForm):
     class Meta:
@@ -22,3 +22,30 @@ class BodyDataForm(ModelForm):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.fields['weight'].widget.attrs.update({'style': 'color: white'})
+
+class FoodForm(ModelForm):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields['name'].widget.attrs.update({'style': 'color: white'})
+        self.fields['calories'].widget.attrs.update({'style': 'color: white'})
+        self.fields['carbs'].widget.attrs.update({'style': 'color: white'})
+        self.fields['protein'].widget.attrs.update({'style': 'color: white'})
+        self.fields['amount'].widget.attrs.update({'style': 'color: white'})
+
+    class Meta:
+        model = Food
+        fields = '__all__'
+
+class ProfileForm(ModelForm):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields['firstname'].widget.attrs.update({'style': 'color: white'})
+        self.fields['lastname'].widget.attrs.update({'style': 'color: white'})
+        self.fields['age'].widget.attrs.update({'style': 'color: white'})
+        self.fields['height'].widget.attrs.update({'style': 'color: white'})
+        self.fields['initWeight'].widget.attrs.update({'style': 'color: white'})
+        self.fields['goalWeight'].widget.attrs.update({'style': 'color: white'})
+
+    class Meta:
+        model = Profile
+        fields = ['firstname', 'lastname', 'age', 'height', 'initWeight', 'goalWeight']
