+++ main_app/templates/base.html
@@ -35,8 +35,8 @@
                                 </form>
                             </li>
                         {% else %}
-                            <li><a href="{% url 'signup' %}">Sign Up</a></li>
-                            <li><a href="{% url 'login' %}">Log In</a></li>
+                            <li><a class="waves-effect waves-light btn-small black" href="{% url 'signup' %}">Sign Up</a></li>
+                            <li><a class="waves-effect waves-light btn-small black" href="{% url 'login' %}">Log In</a></li>
                         {% endif %}
                     </ul>
                 </div>
