+++ main_app/templates/main_app/bodydata_form.html
@@ -12,7 +12,7 @@
     <table>
         {{ form.as_table }}
     </table>
-    <input type="submit" value="Submit!" class="btn blue darken-1 z-depth-3">
+    <input type="submit" value="Submit!" class="btn waves-effect waves-light blue darken-1 z-depth-3">
 </form>
 
 {% endblock %}
