+++ main_app/templates/main_app/food_form.html
@@ -13,11 +13,11 @@
   <table>
     {{ form.as_table }}
   </table>
-  <input type="submit" value="Submit!" class="btn blue darken-1 z-depth-3">
+  <input type="submit" value="Submit!" class="btn waves-effect waves-light blue darken-1 z-depth-3">
 </form>
 <br>
 {% if object %}
-  <a href="{% url 'foods_delete' object.id %}" class="btn blue darken-1 z-depth-3">Delete</a>
+  <a href="{% url 'foods_delete' object.id %}" class="btn waves-effect waves-light blue darken-1 z-depth-3">Delete</a>
 {% endif %}
 
 {% endblock %}
