+++ main_app/templates/main_app/profile_detail.html
@@ -14,7 +14,7 @@
                 <p>Goal Weight: {{ profile.goalWeight }}</p>
             </div>
             <div class="card-action">
-                <a href="{% url 'profile_update' profile.id %}">Edit</a>
+                <a href="{% url 'profile_update' profile.id %}" class="btn waves-effect waves-light blue darken-1 z-depth-3">Edit</a>
             </div>
         </div>
     </div>
